pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "healthsim-agent"
version = "0.1.0"
description = "AI-powered healthcare data simulation agent using Anthropic Agent SDK"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Mark Oswald"}
]
keywords = ["healthcare", "simulation", "synthetic-data", "ai-agent", "anthropic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    # Anthropic Agent SDK
    "anthropic>=0.40.0",
    
    # Database
    "duckdb>=1.1.0",
    
    # CLI and UI
    "rich>=13.9.0",
    "click>=8.1.0",
    "prompt-toolkit>=3.0.0",
    
    # Data validation
    "pydantic>=2.9.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "pathlib>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
]

[project.scripts]
healthsim = "healthsim_agent.main:main"

[project.urls]
Homepage = "https://github.com/mark64oswald/healthsim-agent"
Repository = "https://github.com/mark64oswald/healthsim-agent"
Documentation = "https://github.com/mark64oswald/healthsim-agent#readme"

[tool.hatch.build.targets.wheel]
packages = ["src/healthsim_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
